Make GMSL compatible with .EXPORT_ALL_VARIABLES
This address issue #15

GMSL 1.1.3 was not compatible with .EXPORT_ALL_VARIABLES or a naked
export in a Makefile (both cause all variables to be exported into the
shell used to run commands) because some of the functions defined in
__gmsl would fail if expanded without arguments (i.e. treated as variables
and not functions).

This commit fixes that by protecting those functions that would have failed
and by altering the way int_encode works so there is not a very long string
defined in __gmsl. int_encode now works by building a string on the fly
for any integer to be encoded.

Altered the way the test suite works so that it is run normally and with
.EXPORT_ALL_VARIABLES to check that it still works correctly.

Partial address of issue #14. The tar ball is now made in a directory with
the same name as the gmsl release (e.g. gmsl-1.1.4) and expands into that.

Also, changed the